+++ work-discord-bot/src/api/routes/handleVerifyWallet.ts
@@ -1,9 +1,5 @@
 import { Request, Response } from 'express';
-import { 
-  client, 
-  pendingVerifications, 
-  supabase
-} from '../config';
+import { client, pendingVerifications, supabase } from '../../config';
 import { 
   getServerConfig, 
   verifySignature, 
@@ -12,65 +8,8 @@ import {
   checkAnyWalletHasSufficientBalanceNumber, 
   formatBalanceNumber,
   createOrUpdateHolder
-} from '../utils';
-import { TextChannel, PermissionFlagsBits, GuildMember } from 'discord.js';
-
-export async function handleVerificationContext(req: Request, res: Response) {
-  const verificationCode = req.query.code as string;
-  const logPrefix = `[API Context ${verificationCode?.substring(0, 6)}...]`;
-  console.log(`${logPrefix} Request received.`);
-
-  if (!verificationCode) {
-    console.log(`${logPrefix} Missing verification code`);
-    res.status(400).json({
-      success: false,
-      message: "Missing verification code query parameter.",
-    });
-    return;
-  }
-
-  const verificationData = pendingVerifications.get(verificationCode);
-  if (!verificationData) {
-    console.log(`${logPrefix} Code not found or expired`);
-    res.status(404).json({
-      success: false,
-      message:
-        "Invalid or expired verification code. Please try /verify again.",
-    });
-    return;
-  }
-  const { guildId, userId, action } = verificationData;
-  console.log(
-    `${logPrefix} Found code for user ${userId}, guild ${guildId}, action ${action}`
-  );
-
-  const serverConfig = await getServerConfig(guildId);
-  if (!serverConfig) {
-    console.log(`[API Context] Server config not found for guild: ${guildId}`);
-    res.status(404).json({
-      success: false,
-      message:
-        "Server configuration not found or incomplete for this verification link. Please contact an admin.",
-    });
-    return;
-  }
-  console.log(`[API Context] Found server config for guild: ${guildId}`);
-
-  res.status(200).json({
-    success: true,
-    message: "Verification context retrieved successfully.",
-    context: {
-      userId: userId,
-      action: action,
-      guildId: guildId,
-      tokenAddress: serverConfig.token_address,
-      requiredBalance: serverConfig.required_balance,
-      tokenSymbol: serverConfig.token_symbol,
-      tokenDecimals: serverConfig.token_decimals,
-      serverName: serverConfig.server_name,
-    },
-  });
-}
+} from '../../utils';
+import { PermissionFlagsBits, GuildMember } from 'discord.js';
 
 export async function handleVerifyWallet(req: Request, res: Response) {
   const { verificationCode, walletAddress, signature, message, isLedgerFlow } = req.body;
@@ -352,159 +291,6 @@ export async function handleVerifyWallet(req: Request, res: Response) {
     res.status(500).json({
       success: false,
       message: "A database error occurred while saving wallet information.",
-    });
-    return;
-  }
-}
-
-
-export async function handleSendChannelMessage(req: Request, res: Response) {
-  const { message, channelId, guildId } = req.body;
-  const logPrefix = `[API ChanMsg ${guildId}/${channelId}]`;
-  console.log(`${logPrefix} Request received.`);
-
-  if (!message || !channelId || !guildId) {
-    console.log(`${logPrefix} Missing required fields.`);
-    res.status(400).json({
-      success: false,
-      message: "Missing required fields (message, channelId, guildId).",
-    });
-    return;
-  }
-
-  try {
-    const guild = client.guilds.cache.get(guildId);
-    if (!guild) {
-      console.log(`${logPrefix} Guild ${guildId} not found.`);
-      res.status(404).json({
-        success: false,
-        message: "Bot is not in the specified server.",
-      });
-      return;
-    }
-
-    const channel = await guild.channels.fetch(channelId).catch(() => null);
-
-    if (channel?.isTextBased() && channel instanceof TextChannel) {
-      if (!guild.members.me) {
-        console.error(
-          `${logPrefix} Cannot check permissions as guild.members.me is null.`
-        );
-        res.status(500).json({
-          success: false,
-          message: "Internal bot error checking permissions.",
-        });
-        return;
-      }
-      if (
-        !channel
-          .permissionsFor(guild.members.me)
-          ?.has(PermissionFlagsBits.SendMessages)
-      ) {
-        console.error(`${logPrefix} Bot lacks SendMessages permission.`);
-        res.status(403).json({
-          success: false,
-          message: "Bot lacks permission to send messages in this channel.",
-        });
-        return;
-      }
-
-      await channel.send(String(message)); // Ensure message is string
-      console.log(`${logPrefix} Message sent successfully.`);
-      res.status(200).json({
-        success: true,
-        message: "Message sent successfully!",
-      });
-      return;
-    } else {
-      console.log(`${logPrefix} Channel not found or not a text channel.`);
-      res.status(404).json({
-        success: false,
-        message: "Channel not found or is not a valid text channel.",
-      });
-      return;
-    }
-  } catch (error) {
-    console.error(`${logPrefix} API Error sending channel message:`, error);
-    res.status(500).json({
-      success: false,
-      message: "An internal server error occurred while sending the message.",
-    });
-    return;
-  }
-}
-
-export async function handleSendDirectMessage(req: Request, res: Response) {
-  const { message, userId, guildId } = req.body;
-  const logPrefix = `[API DMsg ${guildId}/${userId}]`;
-  console.log(`${logPrefix} Request received.`);
-
-  if (!message || !userId || !guildId) {
-    console.log(`${logPrefix} Missing required fields.`);
-    res.status(400).json({
-      success: false,
-      message: "Missing required fields (message, userId, guildId).",
-    });
-    return;
-  }
-
-  try {
-    const guild = client.guilds.cache.get(guildId);
-    if (!guild) {
-      console.log(`${logPrefix} Guild ${guildId} not found.`);
-      res.status(404).json({
-        success: false,
-        message: "Bot is not in the specified server context for this request.",
-      });
-      return;
-    }
-
-    const member = await guild.members.fetch(userId).catch(() => null);
-
-    if (member) {
-      try {
-        await member.send(String(message)); // Ensure message is string
-        console.log(`${logPrefix} DM sent successfully.`);
-        res.status(200).json({
-          success: true,
-          message: "Direct message sent successfully!",
-        });
-        return;
-      } catch (dmError: any) {
-        console.error(`${logPrefix} API Error sending DM:`, dmError);
-        if (dmError.code === 50007) {
-          console.log(`${logPrefix} Cannot send DM (user settings or block).`);
-          res.status(403).json({
-            success: false,
-            message:
-              "Cannot send direct message to this user. They may have DMs disabled or have blocked the bot.",
-          });
-          return;
-        } else {
-          res.status(500).json({
-            success: false,
-            message:
-              "An internal server error occurred while attempting to send the direct message.",
-          });
-          return;
-        }
-      }
-    } else {
-      console.log(`${logPrefix} User ${userId} not found in guild ${guildId}.`);
-      res.status(404).json({
-        success: false,
-        message: "User not found in the specified server.",
-      });
-      return;
-    }
-  } catch (error) {
-    console.error(
-      `${logPrefix} API Error (outer scope) processing DM request:`,
-      error
-    );
-    res.status(500).json({
-      success: false,
-      message: "An unexpected internal server error occurred.",
     });
     return;
   }
