+++ work-discord-bot/src/api/routes/index.ts
@@ -1,0 +1,4 @@
+export { handleVerificationContext } from './handleVerificationContext'
+export { handleVerifyWallet } from './handleVerifyWallet';
+export { handleSendChannelMessage } from './handleSendChannelMessage';
+export { handleSendDirectMessage } from './handleSendDirectMessage';
