+++ work-discord-bot/src/bot/commands/index.ts
@@ -1,0 +1,4 @@
+export { handleSetupCommand } from './handleSetupCommand';
+export { handleEditConfigCommand } from './handleEditConfigCommand';
+export { handleVerifyCommand } from './handleVerifyCommand';
+export { handleTipCommand } from './handleTipCommand';
