+++ work-discord-bot/src/bot/interactions/index.ts
@@ -1,0 +1,2 @@
+export { handleButtonInteraction } from './handleButtonInteraction';
+export { handleCommandInteraction } from './handleCommandInteraction';
